Print Methods for r2 and comp_fit calculation Objects

Printing objects of class "r2_kvr2" (generated by r2) or "comp_kvr2" (generated by comp_fit), respectively, by simple print methods.

Usage

# S3 method for class 'r2_kvr2'
print(x, ..., digits = 4)

# S3 method for class 'comp_kvr2'
print(x, ..., digits = 4)

Arguments

x

AN object of class "r2_kvr2" or "comp_kvr2".

...

Further arguments passed to or from other methods.

digits

The number of significant digits to be used for printing. Default to 4.

Details

These methods format the calculated statistics into a human-readable summary, displaying each index or metric with its corresponding value.

See also